fix: prevent duplicate retries on 2xx responses and unrecoverable errors

+++ core/src/main/java/nl/clockwork/ebms/client/delivery/task/DeliveryTaskManagerConfig.java
@@ -51,7 +51,7 @@ public class DeliveryTaskManagerConfig
 	@Value("${deliveryTaskManager.nrAutoRetries}")
 	int nrAutoRetries;
 	@Value("${deliveryTaskManager.autoRetryInterval}")
-	int autoRetryInterval;
+	long autoRetryInterval;
 	@Autowired
 	EbMSDAO ebMSDAO;
